Working camera courtesy of Random Guy and not Corey. Made some fixes to the puzzle room and the spike trap and added the prefab for the puzzle room.

+++ trunk/Assets/Scripts/Environment/Triggers/SpikeTrap.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class SpikeTrap : TrapBase
 {
-	public int travelDist = 0;
+	public float travelDist = 0.0f;
 	private Vector3 startPos = Vector3.zero;
 	public Vector3 travelDir = Vector3.zero;
